+++ nodalia-cards.manifest.js
@@ -1,11 +1,11 @@
 export default {
-  "pkgVersion": "2.2.0-alpha.6",
-  "contentSha256_12": "e94da623529f",
+  "pkgVersion": "2.2.0-alpha.7",
+  "contentSha256_12": "f5abf385d582",
   "file": "nodalia-cards.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
   "editorSha256_12": "fcf4f0088891"
 };
-export const pkgVersion = "2.2.0-alpha.6";
-export const contentSha256_12 = "e94da623529f";
+export const pkgVersion = "2.2.0-alpha.7";
+export const contentSha256_12 = "f5abf385d582";
 export const file = "nodalia-cards.js";